Split compilation
The 'Print_Board function' was declared in main.c
and modified to be defined in print.c.
And the 'Get_Around_Mine_Number' function was declared in main.c
and modified to be defined in game_setting.c

Resolved: #14
