+++ game_setting.h
@@ -2,3 +2,4 @@
 #define BOARD_SIZE 6
 
 int Initialize_Control_Board(char control_board[BOARD_SIZE][BOARD_SIZE], int difficulty);
+int Select_Difficulty();
