+++ print.c
@@ -17,3 +17,29 @@ void Display_Welcome_Message() {
 
     printf("Welcome to Minesweeper. Fork me at: https://github.com/apieceofCAKE/minesweeper_game\n");
 }
+
+
+void Print_Board(char board[BOARD_SIZE][BOARD_SIZE]) {
+    /*
+    Prints the parameter array in a user friendly way
+    */
+
+    printf("\n");
+
+    for (int col = 0; col < BOARD_SIZE; col++) {
+
+        printf("c%d ", col);
+    }
+
+    printf("\n\n");
+
+    for (int row = 0; row < BOARD_SIZE; row++) {
+
+        for (int col = 0; col < BOARD_SIZE; col++) {
+
+            printf(" %c ", board[row][col]);
+        }
+
+        printf("     r%d\n", row);
+    }
+}
