+++ print.h
@@ -1,1 +1,4 @@
+#define BOARD_SIZE 6
+
 void Display_Welcome_Message();
+void Print_Board(char board[BOARD_SIZE][BOARD_SIZE]);
